% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractR.R
\name{extractR}
\alias{extractR}
\title{Extracts mean values from Landsat imagery based on locations provided
by shape files.}
\usage{
extractR(wdir, imdir, option, attrb)
}
\arguments{
\item{wdir}{Character string filepath to the working directory where the
original shape files are located. This working directory must contain a
folder named "site_vectors" which contains the individual shape files for
extraction.}

\item{imdir}{Character string filepath to the path/row level for imagery.}

\item{option}{A character string indicating one of the following "i35",
"ndvi", "b1", "b2", "b3", "b4", "b5", "b6".}

\item{attrb}{A character string of the name of field in the attribute column
of the shape file that contains the unique identifier for the location.
This must be the same as was used to create the jpegs and is usually the site
ID.}
}
\value{
Creates individual .csv files for extracted values per shape file and
another that collates them all.
}
\description{
\code{extractR} will extract mean pixel values from dates, determined from
 quality assessed jpegs, of Landsat raster imagery, from locations as
 provided by shape files. Common usage is for extracting vegetation indices
 from field work or calibration sites. When run it will:
  \enumerate{
      \item Copy the appropriate shape files to the appropriate quality
      assessed jpeg folders based on attribute id.
      \item Looks in each jpeg folder to determine appropriate image dates
      i.e. stores the dates of the remaining jpegs after quality assessment.
      \item Based on user input, extracts mean band or vegetation index
      value per site from the shape file.
      \item Writes mean values to a .csv file and stores each in appropriate
      jpeg folder.
      \item On completion it will join all extracted value data sets into one
      .csv file.
      }
}
\examples{
\dontrun{
extractR(wdir = "z:/working", imdir = "W:/usgs/115078", option = "i35", attrb = "ID")
}

}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}

For more details see  \url{https://rspaw.github.io/RSSApkg/index.html}
{the RSSApkg website}
}
