% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jpegR.R
\name{jpegR}
\alias{jpegR}
\title{Create small jpegs of satellite imagery.}
\usage{
jpegR(wdir, imdir, layer, attrb, start = NA, stop = NA, combo,
  buffer = 2000)
}
\arguments{
\item{wdir}{Character string filepath to the working directory where the
original shape file is located.}

\item{imdir}{Character string filepath to the path/row level for imagery.}

\item{layer}{Character string of the name of the shape file (no extension).}

\item{attrb}{Character string of the name of the field in the attribute
table of the shape file that contains the unique entries (e.g.site numbers).}

\item{start}{Character string representing start date of imagery search. Must
be in the form "dmY" or "dmy" and can include separators. Defaults to NA so
image search starts from first available image in the RSSA archive.}

\item{stop}{Character string representing end date of imagery search. Must
be in the form "dmY" or "dmy" and can include separators. Defaults to NA so
image search ends with latest available image in the RSSA archive.}

\item{combo}{Numerics representing bands to place in RGB channels of the jpeg.
Must be of the form c(N\emph{r}, N\emph{g}, N\emph{b}) where N\emph{r} is the
band number for the red channel etc.}

\item{buffer}{Numeric representing how much to buffer out the site to control
zoom level for jpegs. If imagery crs is projected, number represents metres.
If imagery is geodetic then the number represents degrees.}
}
\value{
Creates named folders, based on site location and dates of images
processed, containing small jpeg images for visual assessment.
}
\description{
\code{jpegR} makes small jpeg images centered on and showing site locations
from a shape file. This function can be used as part of the cloud quality
check process prior to extracting satellite bands or indices. Although
designed to help the QA process the jpegs can be useful for quickly checking
many image dates for change (e.g. fires, vegetation removal, etc) or for
creating stills for animating. When run the function:
\enumerate{
    \item Obtains a list of imagery folders to access. In doing so it also
    checks for folder name errors (\code{\link{u_leapR}}) and only returns
    sun-corrected image folders.
    \item Splits a supplied shape file by unique attribute field entries,
    stores results in a folder and writes a text file containing the original
    shape file name (\code{\link{u_shpsplitR}}).
    \item Creates a small jpeg of the individual sites for each image date as
    per step 1. jpegs are based on band combination from user input and
    buffered by a desired distance.
    \item All outputs are placed in appropriately named folders per site in
    the working directory
    }
}
\examples{
\dontrun{
jpegR(wdir = "z:/working", imdir = "W:/usgs/115078", layer = "site_locations",
attrb = "ID", start = "1/1/98", stop = "28-12-2009", combo = c(5,4,2),
buffer = 2000)
}

}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}

For more details see  \url{https://bartesto.github.io/RSSApkg/index.html}
{the RSSApkg website}
}
