% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSA_utils.R
\name{u_dateR}
\alias{u_dateR}
\title{Creates a data frame of paths, folder names and dates.}
\usage{
u_dateR(path, archive, pat = ".jpeg")
}
\arguments{
\item{path}{Character string filepath to the path/row of satellite imagery to
check. Can also be a folder of jpegs or some other date named files.}

\item{archive}{A logical scalar. \strong{TRUE}  will assume \strong{path} is
indicating RSSA internal USGS archive. \strong{FALSE} can be used when
querying a QA jpeg folder or the like.}

\item{pat}{Pattern to search for in a non-USGS imagery archive. Defaults to
".jpeg".}
}
\value{
For a USGS imagery archive the query will return a data frame with:
\itemize{
    \item path - a column of character string file paths
    \item folds - a column of character string folder names
    \item dates - a column of date class dates
    }
For a non-USGS archive the query will return a data frame with:
\itemize{
    \item path - a column of character string file paths
    \item folds - a column of character string folder names
    }
}
\description{
\code{u_dateR} creates a data frame of paths, folder names and dates from the
contents of a folder.
}
\details{
This output of this function will hold paths, folder names and dates that will
be used for other functions such as \code{jpegR} and \code{extractR}.
}
\examples{
\dontrun{
u_dateR(path = "W:/usgs/115078", archive = TRUE)
u_dateR(path = "Z:/DEC/working/QA...", archive = FALSE, pat = ".jpeg")
}

}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
