% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSA_utils.R
\name{u_shpsplitR}
\alias{u_shpsplitR}
\title{A tool for splitting ESRI shape files}
\usage{
u_shpsplitR(pathin = ".", pathout = ".", shp, shp.ID)
}
\arguments{
\item{pathin}{Character string filepath to the location of the shape file.}

\item{pathout}{Character string filepath to the write location.}

\item{shp}{Character string of the name of the shape file (no extension).}

\item{shp.ID}{Character string of the name of the field in the attribute
table of the shape file that contains the unique entries (e.g.site numbers).}
}
\value{
Creates multiple ESRI shape files each named after the unique entry
found in 'shp.ID'.
}
\description{
\code{u_shpsplitR} takes an ESRI shape file and splits it into multiple shape
files based on unique entries in a field of the attribute table.
}
\examples{
\dontrun{
u_shpsplitR("plot_locations", "Plot_ID")
}


}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
