% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSA_utils.R
\name{u_leapR}
\alias{u_leapR}
\title{Renames incorrectly named leap year image date folders.}
\usage{
u_leapR(path)
}
\arguments{
\item{path}{Character string filepath to the path/row of imagery to check.}
}
\value{
If incorrectly named folders are identified they are renamed a day
    earlier, otherwise a message "No leap date folder errors" is printed to
    screen.
}
\description{
\code{u_leapR} reads the date of a suncorrected image, checks it against the
folder name and if incorrect, renames the folder. This corrects for leap year
folder naming errors propagated by batch processing.
}
\examples{
\dontrun{
u_leapR("W:/usgs/115078")
}

}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
