% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSA_utils.R
\name{u_shpsplitR}
\alias{u_shpsplitR}
\title{A tool for splitting ESRI shape files.}
\usage{
u_shpsplitR(pathin, pathout = paste0(pathin, "/site_vectors"), layer, attrb)
}
\arguments{
\item{pathin}{Character string filepath to the location of the shape file.}

\item{pathout}{Character string filepath to the write location.}

\item{layer}{Character string of the name of the shape file (no extension).}

\item{attrb}{Character string of the name of the field in the attribute
table of the shape file that contains the unique entries (e.g.site numbers).}
}
\value{
Creates multiple ESRI shape files each named after the unique entry
found in 'attrb' and saves them to a folder (defaults to "site_vectors"). It
will also write a text file with the name of the original layer file used and
output names of new layers created to screen or to object if created.
}
\description{
\code{u_shpsplitR} takes an ESRI shape file and splits it into multiple shape
files based on unique entries in a field of the attribute table.
}
\examples{
\dontrun{
u_shpsplitR(layer = "plot_locations", attrb = "Plot_ID")
}

}
\author{
Bart Huntley, \email{bart.huntley@dpaw.wa.gov.au}
}
